test: add test to the list implementation
